# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Moco API",
    "description": "Moco API 문서입니다.",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/admin/createAdmin": {
      "post": {
        "tags": ["관리자 기능"],
        "summary": "관리자 생성",
        "description": "role = 1 슈퍼관리자만 사용 가능",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "string"
                },
                "phone": {
                  "example": "string"
                },
                "email": {
                  "example": "string"
                },
                "password": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "관리자가 성공적으로 생성되었습니다.",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "필수 항목을 모두 입력해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "권한이 없습니다.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "관리자 생성 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/statisticsDetail": {
      "get": {
        "tags": ["관리자 기능"],
        "summary": "관리자 테스트 상세 통계",
        "description": "해당 테스트(id) 의 통계(count) 를 조회 합니다. totalCount (총 이용횟수) 가 같이 내려갑니다.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "testId",
            "in": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 상세 통계 조회를 성공적으로 가져왔습니다.",
                  "type": "string"
                },
                "result": {
                  "example": {
                    "resultData": [
                      {
                        "id": "number",
                        "test_id": "number",
                        "type": "string",
                        "description": "string",
                        "count": "number"
                      }
                    ],
                    "totalCount": "number"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "유효하지 않은 토큰 입니다.",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "권한이 없습니다..",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 상세 통계 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/statisticsList": {
      "get": {
        "tags": ["관리자 기능"],
        "summary": "관리자 테스트 통계 목록",
        "description": "테스트 목록 중 totalCount 가 1 이상인 목록을 조회 합니다.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 통계 목록을 성공적으로 가져왔습니다..",
                  "type": "string"
                },
                "result": {
                  "example": {
                    "result": {
                      "id": "number",
                      "name": "string",
                      "path": "string",
                      "totalCount": "number"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "유효하지 않은 토큰 입니다.",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "권한이 없습니다..",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 통계 목록 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/login": {
      "post": {
        "tags": ["관리자 인증"],
        "summary": "인증",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "string"
                },
                "password": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "로그인 성공",
                  "type": "string"
                },
                "result": {
                  "example": {
                    "accessToken": "string",
                    "refreshToken": "string",
                    "role": "number",
                    "username": "string",
                    "email": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "필수 항목을 모두 입력해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "사용자를 찾을 수 없습니다.",
                  "type": "string"
                }
              }
            }
          },
          "402": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "비밀번호가 일치하지 않습니다.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "로그인 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/admin/logout": {
      "post": {
        "tags": ["관리자 인증"],
        "summary": "로그아웃",
        "description": "로그아웃",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/admin/refreshToken": {
      "post": {
        "tags": ["관리자 인증"],
        "summary": "토큰 재발급",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "토큰이 성공적으로 재발급되었습니다.",
                  "type": "string"
                },
                "result": {
                  "example": {
                    "accessToken": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "리프레시 토큰을 제공해야 합니다.",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "리프레시 토큰이 유효하지 않습니다.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "토큰 재발급 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}